requirements.txt
cryptography>=41.0.0
argon2-cffi>=23.1.0
click>=8.1.0
rich>=13.0.0